+++ tests/resources/do_db_ddl_clean.sql
